refactor: evaluationLogicからstaffConstraintAnalyzerとsimulationGeneratorを抽出
- analyzeStaffConstraints()をstaffConstraintAnalyzer.tsに抽出（198行）
- generateSimulation()をsimulationGenerator.tsに抽出（84行）
- evaluationLogic.tsを588行から352行に削減（-40%）
- 全213件のユニットテストがパス
